+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### 2026-2-26
+
+#### New:
+- 加入素数判断[`NumberTheory::IsPrime::miller_rabin`]
+
+-----
 ### 2026-2-8
 
 #### New: